feat(node): expose UUID v7 and cookie helpers
Lifts the cookie helpers from @posthog/next to @posthog/core and re-exports
them from posthog-node, so consumers managing distinct_id outside the
browser SDK (e.g. Lambda functions handing out cross-domain redirects) can
read/write the PostHog cookie format without reverse-engineering it.

@posthog/next now re-exports the same surface from @posthog/core to keep
existing internal imports working without churn.

Closes #2143
