+++ packages/next/src/shared/constants.ts
@@ -1,5 +1,3 @@
-export const COOKIE_PREFIX = 'ph_'
-export const COOKIE_SUFFIX = '_posthog'
 export const DEFAULT_API_HOST = 'https://us.i.posthog.com'
 export const COOKIE_MAX_AGE_SECONDS = 365 * 24 * 60 * 60
 export const DEFAULT_INGEST_PATH = '/ingest'
