+++ packages/node/src/exports.ts
@@ -8,3 +8,17 @@ export { FeatureFlagEvaluations } from './feature-flag-evaluations'
 // These were originally defined in posthog-node and moved to core for reuse across SDKs.
 export { FeatureFlagError } from '@posthog/core'
 export type { FeatureFlagErrorType } from '@posthog/core'
+
+// Identity helpers re-exported from core for posthog-node consumers managing
+// distinct_id outside the browser SDK (e.g. Lambda functions handing out
+// `download-app` redirects). Closes #2143.
+export {
+  cookieStateToProperties,
+  cookieStoreFromHeader,
+  getPostHogCookieName,
+  parsePostHogCookie,
+  readPostHogCookie,
+  serializePostHogCookie,
+  uuidv7,
+} from '@posthog/core'
+export type { CookieStore, PostHogCookieState } from '@posthog/core'
